#4 - File cleanup and documenting new types and functions
