--- database/sql_adapter/sql_script_repository.go
@@ -1,5 +1,0 @@
-package sqladapter
-
-type SQLScriptRepository interface {
-	GetSQLScript(key string) string
-}
